perf: implement data key caching to improve request throughput
Currently we generate a new data key for every save to the session,
which incurs a calls to KMS. Additionally, every read from the session
calls KMS.

Here we use the AWS Encryption SDK to implement data key caching to
reduce calls to KMS. Currently the configuration for this is defaulted
to the following values:

Max no. keys held in the cache: 100
Max age of each key: 30 seconds
Max no. messages encrypted using the same key: 100

Resolves AFPC-72

GitOrigin-RevId: 4305e9314ebc28a4fcae18607f1823f83b817423