feat: remove any GA cookies if policy is rejected after it was accepted
When a user accepts the cookie prompt the Google Tag Manager snippet is
added to the page which then sets tracking cookies. If the user then
rejects cookies the snippet is removed but the cookies linger in the
browser - this change actively removes any remaining cookies when the
cookie policy is rejected.

Resolves AFPC-648

GitOrigin-RevId: 2e4cff0d0c62aee345f1a5b80a8d2614f0e5eca3