feat: update money page to include new content and money specific input
We have users getting validation errors on the money-you-have-page, some
of which fail to progress to the next page and drop out of the service.
These changes make it clearer we want a total numeric figure, with
content tweaks to the field labels, hints and error messages. We're
also making use of GOV.UK's new input prefixes to include a pound sign
in the field itself.

Resolves AFPC-811

GitOrigin-RevId: 92b0354349930b26bd3a7266e4ca6d02847af7c7