feat: persist initial application date from start of journey to the end
If a claimant starts their claim before midnight but submits it after
the submitted dates of claim may differ from the ones they were shown
in the service. To fix this we establish a fixed 'application date' when
the claim is started and all date calculations are made from that one
date.

GitOrigin-RevId: 5dca292ef657efbe1ed2345afdcedba0b7490b28