feat: [bc] offer calculated claim date to user rather than take any date
BREAKING CHANGE: New pages added, one page removed. Journey's changed
in such a way that would required user to go back and answer unanswered
questions.

Claimants don't understand the guidance on the Date of Claim page and
are entering invalid dates of claim, or potentially missing out on
backdating entitlement. To resolve this we are calculating a date of
claim in service based on their date of birth and any periods abroad and
are offering it to them as a Yes/No option. They can still supply their
own date of claim if they wish.

Resolves AFPC-180, AFPC-787

GitOrigin-RevId: 4641f5554e1a6dcba93215d50169cfab019fca54