refactor: move CYA print button outside of form tag so it cannot submit
In the event the client side javascript doesn't load, fails, or the
preventDefault method for some reason does not work or is not supported,
the print button will submit the form on the CYA page. This is because
the button sits within the form tag and submitting the form is its
default on-click behaviour. This would be unexpected behaviour for a
user and could lead them to submitting their answers unintentionally. To
avoid this I'm moving the print button outside of the form tag.

GitOrigin-RevId: f35986e4f1e896ba41ad4e90b7b8cf015f071046