fix: sanitise money string with commas prior to calculation
This fix resolves an issue with the money-you-have and money-today
values that the claimaint enter. If they entered a value with a comma
(e.g. 10,001.00), post formatting the number value would be 10. The fix
checks for the presence of , and . and strips them before parsing the number
so 10,001.00 becomes 10001. This figure is then accurate for the follow on
calculation for disregards.

Resolves AFPC-956

GitOrigin-RevId: 50dbfc62b47c2625a9d99040f2c37b3a9c7fe42c