feat: [bc] ask home owners if they have a mortgage
Previously we were asking all home owners if they wanted support with
interest on a mortgage or if they shared the mortgage with someone else.
This meant that we were also asking this of claimants who had already
paid their mortgages off, leading to confusion. Now we check if a
claimant has a mortgage before asking more mortgage related questions.

BREAKING CHANGE: This adds an extra required page into the journey, so
would break any active sessions that are past the point this question
should have been asked. CASA would force the user back to the mortgage
page in this instance.

Resolves AFPC-781

GitOrigin-RevId: 33ea7c03227b5a013d26fa56d393a98cc2132a52